Added compat tolerance for commas
Previously, user-provided matrix string representations
were only allowed to use spaces as the delimiter between
successive entries in the same row of a matrix.
This commit allows commas to be used, too.
Note that it is not acceptable to mix commas and spaces
as delimiters on the same row. If there is a comma
anywhere in the row, commas will be treated as delimiters;
otherwise, spaces will be treated as delimiters.

This better matches the MATLAB matrix syntax, making it
easier to go back and forth between C++ and MATLAB.